requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.2.0
autobahn==24.4.2
Automat==24.8.1
bidict==0.23.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
distro==1.9.0
Django==5.1.6
django-cors-headers==4.7.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jiter==0.8.2
msgpack==1.1.0
openai==1.64.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
pyOpenSSL==25.0.0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
redis==5.2.1
requests==2.32.3
service-identity==24.2.0
setuptools==75.8.0
simple-websocket==1.1.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wheel==0.45.1
wsproto==1.2.0
zope.interface==7.2
